+++ src/pykdex/estimators/__init__.py
@@ -3,8 +3,14 @@
 
 """User-facing density and intensity estimators."""
 
+from pykdex.estimators.heat_experiment import HeatNetworkExperiment
 from pykdex.estimators.heat_network_kde import HeatNetworkKDE
 from pykdex.estimators.network_kde import NetworkKDE
 from pykdex.estimators.spatial_kde import SpatialKDE
 
-__all__ = ["SpatialKDE", "NetworkKDE", "HeatNetworkKDE"]
+__all__ = [
+    "SpatialKDE",
+    "NetworkKDE",
+    "HeatNetworkKDE",
+    "HeatNetworkExperiment",
+]